Add files via upload
The front and back of the unit showing the proto board, I used silicon wires as they are easier to solder and not make a mess with the tight wiring. The 7805 regulator is on the other side of the board and has a .47uf capacitor on the input side.   I used a little 3v3 regulator as well so both the 7805 and the 3.3v regulator are fed from a typical 9-12 volt supply.  The power supply options are not shown on the wiring schedule.