Create siggenv1.0.ino
There is a lot of room for improvement and  probably is missing credits to someone as this world is a bit incestuous it is hard to keep up with things end result is quite satisfying as it is an instrument I actually need and works way past frequencies I actually use. I want to work on a signal analysis feature which cannot fit on this hardware so I need to box this up and move onto my next project - ideally I will integrate the signal generator with my next project as the sig gen is not processor intensive it is just a memory challenge however the other project needs a bit more coal and steam than I can get from lower end arduino tech. I did not mention much about the piezo buzzer but hopefully it is self explanatory.