Add files via upload
you can buy a cable from the makers of the MASSDUINO core however i forgot to order in time so make one up.. 1, took standard header pins and filed the plastic down to make them fit the smaller connector, the pins themselves are fine but the header is a bit smaller, clearly if you have a smaller pitch header you will be better off. I soldered directly to the serial adapter board which are cheap to buy from ebay or aliexpress etc.  Bitsbox carry some items in the UK so worth a try also. The wiring is point to point. I used a hot glue gun to join the header together and some heatshrink to keep it all neat.  FT322RL is the thing to search for if you want to make your own. If you use a larger UNO or Nano you will not need this.